incorporate feedback from Steven and Ryan
